fix
git-svn-id: https://ng.gnunet.org/svn/GNUnet@9659 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/util/crypto/hostkeytest.c
@@ -73,7 +73,7 @@ testEncryptDecrypt ()
       if (strncmp (TESTSTRING, result, strlen (TESTSTRING)) != 0)
         {
           printf ("%s != %.*s - testEncryptDecrypt failed!\n",
-                  TESTSTRING, MAX_TESTVAL, result);
+                  TESTSTRING, (int) MAX_TESTVAL, result);
           ok++;
           continue;
         }
